fix: move check consistency function to VD module
