+++ deps/wazuh_testing/wazuh_testing/end_to_end/vulnerability_detector.py
@@ -8,6 +8,8 @@ Functions:
     - check_vuln_state_index: Check vulnerability state index for a host.
     - get_indexed_vulnerabilities_by_agent: Get indexed vulnerabilities by agent.
     - check_vuln_alert_indexer: Check vulnerability alerts in the indexer for a host.
+    - parse_vulnerability_detector_alerts: Parse vulnerability detector alerts.
+    - check_vuln_state_consistency: Check vulnerability state consistency.
 
 
 Copyright (C) 2015, Wazuh Inc.
@@ -204,3 +206,81 @@ def check_vuln_alert_indexer(vulnerabilities_alerts: Dict, host: str, package: D
                                               'PACKAGE_VERSION': package_version})
 
     return expected_alerts_not_found
+
+
+def check_vuln_state_consistency(vulnerabilities_alerts, vulnerabilities_states):
+    # Get the indexer values
+    alerts_vulnerabilities = {}
+    indices_vulnerabilities = {}
+
+    for agent, vuln_alerts in vulnerabilities_alerts.items():
+        for vuln_alert in vuln_alerts:
+            if agent != vuln_alert['_source']['agent']['name']:
+                logging.error("Agent name is not the same as the agent in the alert")
+            alert_agent = vuln_alert['_source']['agent']['name']
+            alert_cve = vuln_alert['_source']['data']['vulnerability']['cve']
+            alert_package_version = vuln_alert['_source']['data']['vulnerability']['package']['version']
+            alert_package_name = vuln_alert['_source']['data']['vulnerability']['package']['name']
+
+            if agent not in alerts_vulnerabilities:
+                alerts_vulnerabilities[agent] = []
+
+            alerts_vulnerabilities[agent].append({
+                'cve': alert_cve,
+                'agent': alert_agent,
+                'package_name': alert_package_name,
+                'package_version': alert_package_version
+            })
+
+    for agent, vulnerabilities in vulnerabilities_states.items():
+        for vulnerability in vulnerabilities:
+            if agent != vulnerability['_source']['agent']['name']:
+                logging.critical("Agent name is not the same as the agent in the vulnerability state")
+
+            state_agent = vulnerability['_source']['agent']['name']
+            state_cve = vulnerability['_source']['vulnerability']['id']
+            state_package_name = vulnerability['_source']['package']['name']
+            state_package_version = vulnerability['_source']['package']['version']
+
+            if agent not in indices_vulnerabilities:
+                indices_vulnerabilities[agent] = []
+
+            indices_vulnerabilities[agent].append({
+                'cve': state_cve,
+                'agent': state_agent,
+                'package_name': state_package_name,
+                'package_version': state_package_version
+            })
+
+    if vulnerabilities_states.keys() != vulnerabilities_alerts.keys():
+        logging.critical("The number of agents is not the same between alerts and states")
+
+    agents_in_alerts_states = [agent for agent in vulnerabilities_states.keys() \
+            if agent in vulnerabilities_alerts.keys()]
+
+    alerts_not_in_states = []
+    states_not_in_alerts = []
+
+    for agent in agents_in_alerts_states:
+        agent_alerts = [] if agent not in alerts_vulnerabilities else alerts_vulnerabilities[agent]
+        agent_states = [] if agent not in indices_vulnerabilities else indices_vulnerabilities[agent]
+
+        if len(agent_alerts) != len(agent_states):
+            logging.critical(f"The number of alerts is not the same as the number of states for agent {agent}")
+            logging.critical(f"Alerts: {len(agent_alerts)}")
+            logging.critical(f"States: {len(agent_states)}")
+
+        # Check that all alerts are in the index
+        for alert in agent_alerts:
+            if alert not in agent_states:
+                alerts_not_in_states.append(alert)
+
+        # Check that all index states are in the alerts
+        for state in agent_states:
+            if state not in agent_alerts:
+                alerts_not_in_states.append(state)
+
+    return {
+        'alerts_not_in_states': alerts_not_in_states,
+        'states_not_in_alerts': states_not_in_alerts
+        }
